+++ src/index.ts
@@ -1,14 +1,8 @@
-import domToStencil from "./dom-to-stencil";
-// import htmlToDOM from "html-dom-parser/lib/html-to-dom-client";
-import { h, VNode } from "@stencil/core";
-import { HTMLDOMParser } from "./utilities";
+import { VNode } from "@stencil/core";
+import { domToStencil } from "./domToStencil";
+import { domparser } from "./domparser";
 
-export function parse(html: string): (VNode|string)[] {
-  const domNodes = HTMLDOMParser(html);
-  const stencilNodes = domToStencil(domNodes);
-  // @ts-ignore
-  // return [h("div", {})];
-  return stencilNodes;
+export function parse(html: string): (VNode | string)[] {
+  const domNodes = domparser(html);
+  return domToStencil(domNodes);
 }
-
-
